This fixed:2206 and fixed:2208
There seems to be 1 line that went missing in the configurations that I did.

Credits and thanks:
Thanks to faoiseamh, and Biohaz4rd for sending me their config files so I can do some comparisons, this turned out to be not the case.
Thanks to polbat02, and Nitrozik for going in game and testing to find out what revision started this issue.
Finally, thanks to RaZer (aka APOC), for doing about a thousand tests as we pinpointed and solved this MAJOR issue :).